Refactor: Refactor Main component
* Integrate unnecessary components

* Rename Styles file
