+++ src/components/Main/Card/Card.styles.ts
