+++ src/components/Main/Card/Card.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import SVGIconBtn from '~/components/SVGIconBtn/SVGIconBtn';
 import * as colors from '~/styles/common/colors';
 import * as sizes from '~/styles/common/sizes';
-import { InnerBtnWrapper, StyledCard, StyledCardProps } from './StyledCard';
+import { InnerBtnWrapper, StyledCard, StyledCardProps } from './Card.styles';
 import { IQuoteData } from '~/types/interface';
 import { DownloadIcon, EmptyHeartIcon, FillHeartIcon } from '~/assets';
