+++ src/components/Main/ChevronBox/ChevronBox.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ChevronButton, StyledChevronWrapper } from './StyledChevronWrapper';
+import { ChevronButton, ChevronWrapper } from './ChevronBox.styles';
 import * as colors from '~/styles/common/colors';
 import { debounce } from '~/utils/debounce';
 import { LeftChevron, RightChevron } from '~/assets';
@@ -8,19 +8,19 @@ interface IChevronWrapper {
     requestData(id?: string): Promise<any>;
 }
 
-const ChevronWrapper = ({ requestData }: IChevronWrapper) => {
+const ChevronBox = ({ requestData }: IChevronWrapper) => {
     const onClick = debounce(requestData, 300);
 
     return (
-        <StyledChevronWrapper>
+        <ChevronWrapper>
             <ChevronButton direction={"prev"} data-role="prev" onClick={() => onClick()}>
                 <LeftChevron fill={colors.MAIN_WHITE} />
             </ChevronButton>
             <ChevronButton direction={"next"} data-role="next" onClick={() => onClick()}>
                 <RightChevron fill={colors.MAIN_WHITE} />
             </ChevronButton>
-        </StyledChevronWrapper>
+        </ChevronWrapper>
     );
 };
 
-export default ChevronWrapper;
+export default ChevronBox;
