+++ src/components/Main/Main.tsx
@@ -1,9 +1,11 @@
 import React, { MutableRefObject } from 'react';
 import Nav from '~/components/Nav/Nav';
-import Section from '~/components/Main/Section/Section';
-import { StyledMain } from './StyledMain';
+import Card from './Card/Card';
+import ChevronBox from './ChevronBox/ChevronBox';
+import { CardWrapper, StyledMain } from './Main.styles';
 import { IQuoteData, IQuotesState, IUserInfo } from '~/types/interface';
 import { CardPositionType, ChevronEventType } from '~/types/type';
+import { downloadToImg } from '~/services/html2canvas';
 
 interface IProps {
     cardWrapperRef: MutableRefObject<HTMLDivElement | null>;
@@ -23,16 +25,34 @@ interface IProps {
     onClickNavContent(target: IQuoteData): void;
 }
 
+const cardPositions: CardPositionType[] = ['front', 'back'];
+
 const Main = ({ cardWrapperRef, isNavOpen, displayQuotes, historyList, cardPosition, favoriteList, userInfo, isLoggedIn, requestData, handleNav, handleLoginBox, onLogout, onChangeFavorite, onClickNavContent }: IProps) => {
+    const onDownload = (): void => {
+        if(cardWrapperRef.current) {
+            downloadToImg(cardWrapperRef.current);
+        }
+    }
+
     return(
         <StyledMain>
-            <Section 
-                cardWrapperRef={cardWrapperRef}
-                displayQuotes={displayQuotes}
-                cardPosition={cardPosition}
-                requestData={requestData}
-                onChangeFavorite={onChangeFavorite}
-            />
+            <section className='section'>
+                <div className='card_section'>
+                    <CardWrapper ref={cardWrapperRef}>
+                        {cardPositions.map((position, i) => (
+                            <Card 
+                                key={i}
+                                position={position} 
+                                cardPosition={cardPosition}
+                                quoteContent={displayQuotes[position]}
+                                onDownload={onDownload}
+                                onChangeFavorite={onChangeFavorite}
+                            />
+                        ))}
+                    </CardWrapper>
+                    <ChevronBox requestData={requestData} />
+                </div>
+            </section>
             <Nav 
                 isNavOpen={isNavOpen} 
                 historyList={historyList}
