--- src/components/Main/Section/CardWrapper/CardWrapper.tsx
@@ -1,35 +1,0 @@
-import React, { forwardRef } from 'react';
-import { IQuoteData, IQuotesState } from '~/types/interface';
-import { CardPositionType } from '~/types/type';
-import Card from '~/components/Main/Section/CardWrapper/Card';
-import { StyledCardWrapper } from './StyledCardWrapper';
-
-interface IProps {
-    cardPosition: CardPositionType;
-    displayQuotes: IQuotesState;
-    onDownload(): void;
-    onChangeFavorite(target: IQuoteData): void;
-}
-
-const CardWrapper = forwardRef<HTMLDivElement | null, IProps>(({ cardPosition, displayQuotes, onDownload, onChangeFavorite }, ref) => {
-    return (
-        <StyledCardWrapper ref={ref}>
-            <Card 
-                position={"front"} 
-                cardPosition={cardPosition}
-                quoteContent={displayQuotes.front}
-                onDownload={onDownload}
-                onChangeFavorite={onChangeFavorite}
-            />
-            <Card 
-                position={"back"} 
-                cardPosition={cardPosition}
-                quoteContent={displayQuotes.back}
-                onDownload={onDownload}
-                onChangeFavorite={onChangeFavorite}
-            />
-        </StyledCardWrapper>
-    );
-});
-
-export default CardWrapper;
