--- src/components/Main/Section/CardWrapper/StyledCardWrapper.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const StyledCardWrapper = styled.div`
-    position: relative;
-    width: 100%;
-    height: 100%;
-    transform-style: preserve-3d;
-    transition: transform 0.6s ease-in-out;
-`;
